Move test for older Gradle versions to appropriate directory
